filter_rewrite: Changed Teleflask to BotServer, and TeleflaskBase to be now Teleflask
